refactor db.go removing the  db init
